+++ CHANGELOG.md
@@ -7,6 +7,9 @@
   * Added system-wide registry.
   * Added ExshomeDependency.dependency_module/1 function.
   * Use SystemRegistry for TestRegistry.
+  * Add default timeout to the GenServerDependency call function. Running tests with a `--trace` flag will allow you to debug your code without interruptions.
+  * Allow to set value for all variables.
+  * Add support for readonly variables.
 
 ## v0.1.7 (2022-06-28)
   * Created database if it does not exist before migraions.